% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMallowsSMC2-package.R
\docType{package}
\name{BayesMallowsSMC2-package}
\alias{BayesMallowsSMC2}
\alias{BayesMallowsSMC2-package}
\title{BayesMallowsSMC2: Nested Sequential Monte Carlo for the Bayesian Mallows Model}
\description{
Estimation of Mallows' rank model using the SMC2 algorithm (Chopin et al. (2013) \doi{10.1111/j.1467-9868.2012.01046.x}). Supports data in the form of complete and partial rankings, and pairwise preferences.
}
\keyword{internal}
