% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complete_rankings_mixtures}
\alias{complete_rankings_mixtures}
\title{Simulated Data}
\format{
\subsection{\code{complete_rankings_mixtures}}{

A data frame with 1000 rows and 7 columns:
\describe{
\item{timepoint}{Timepoint}
\item{user}{User id}
\item{item1, item2, item3, item4, item5}{Ranking given to the item.}
}
}
}
\usage{
complete_rankings_mixtures
}
\description{
Simulated Data
}
\keyword{datasets}
