% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairwise_preferences_heterogeneous}
\alias{pairwise_preferences_heterogeneous}
\title{Simulated Data with Pairwise Preferences}
\format{
\subsection{\code{pairwise_preferences_heterogeneous}}{

A data frame with 800 rows and 4 columns:
\describe{
\item{timepoint}{Timepoint}
\item{user}{User id}
\item{top_item}{Item preferred in the comparison.}
\item{bottom_item}{Item disfavored in the comparison.}
}
}
}
\usage{
pairwise_preferences_heterogeneous
}
\description{
Simulated Data with Pairwise Preferences
}
\keyword{datasets}
