% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairwise_preferences_mixtures}
\alias{pairwise_preferences_mixtures}
\title{Simulated Data with Pairwise Preferences}
\format{
\subsection{\code{pairwise_preferences_mixtures}}{

A data frame with 4000 rows and 4 columns:
\describe{
\item{timepoint}{Timepoint}
\item{user}{User id}
\item{top_item}{Item preferred in the comparison.}
\item{bottom_item}{Item disfavored in the comparison.}
}
}
}
\usage{
pairwise_preferences_mixtures
}
\description{
Simulated Data with Pairwise Preferences
}
\keyword{datasets}
