% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partial_rankings}
\alias{partial_rankings}
\title{Simulated data with top-3 rankings}
\format{
A data frame with 100 rows and 12 columns:
\describe{
\item{timepoint}{Integer specifying the timepoint.}
\item{user}{Integer specifying the user id.}
\item{item1, item2,...}{Ranking given to each item.}
}
}
\usage{
partial_rankings
}
\description{
A simulated dataset containing top-3 rankings of ten items by 100 users,
at 100 separate timepoints. The dataset was simulated using the
\code{sample_mallows} function of the \code{BayesMallows} package
\insertCite{sorensenBayesMallowsPackageBayesian2020}{BayesMallowsSMC2}, and
is identical to \link{complete_rankings} with rankings below 3 set to \code{NA}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
