% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partial_rankings_mixtures}
\alias{partial_rankings_mixtures}
\title{Simulated Data with Missing Values}
\format{
\subsection{\code{partial_rankings_mixtures}}{

A data frame with 1000 rows and 7 columns:
\describe{
\item{timepoint}{Timepoint}
\item{user}{User id}
\item{item1, item2, item3, item4, item5}{Ranking given to the item.}
}
}
}
\usage{
partial_rankings_mixtures
}
\description{
Simulated Data with Missing Values
}
\keyword{datasets}
