% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.BayesMallowsSMC2}
\alias{print.summary.BayesMallowsSMC2}
\title{Print Method for summary.BayesMallowsSMC2 Objects}
\usage{
\method{print}{summary.BayesMallowsSMC2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.BayesMallowsSMC2}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints a summary of a BayesMallowsSMC2 model.
}
