% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_smc}
\alias{run_smc}
\title{Run Sequential Monte Carlo Algorithm (Internal Function)}
\usage{
run_smc(
  input_timeseries,
  input_prior,
  input_options,
  input_sort_matrices,
  input_sort_counts
)
}
\arguments{
\item{input_timeseries}{List of time series data.}

\item{input_prior}{List of prior parameters.}

\item{input_options}{List of algorithm options.}

\item{input_sort_matrices}{List of precomputed sort matrices.}

\item{input_sort_counts}{List of precomputed sort counts.}
}
\value{
List containing SMC algorithm results.
}
\description{
This is an internal function that implements the core SMC² algorithm.
Users should typically use \code{\link{compute_sequentially}} instead.
}
\keyword{internal}
