% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior parameters for Bayesian Mallows model}
\usage{
set_priors(alpha_shape = 1, alpha_rate = 0.5, cluster_concentration = 10)
}
\arguments{
\item{alpha_shape}{Shape parameter of gamma prior for \code{alpha}.}

\item{alpha_rate}{Rate parameter of gamma prior for \code{alpha}.}

\item{cluster_concentration}{Concentration parameter of Dirichlet prior for
\code{tau}.}
}
\value{
An object to be provided to the \code{priors} argument of
\code{\link[=compute_sequentially]{compute_sequentially()}}.
}
\description{
This function is used to set the parameters of the prior distributions for
the static parameters. For the dispersion parameter \code{alpha} we use a gamma
prior, specified with \code{alpha_shape} and \code{alpha_rate}. For the Dirichlet
process mixture, the parameter is \code{cluster_concentration}. For the modal
ranking we always use a uniform prior, and this is hence not specified.
}
