% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_rank_data.R
\name{setup_rank_data}
\alias{setup_rank_data}
\title{Setup rank data for sequential estimation}
\usage{
setup_rank_data(rankings, preferences)
}
\arguments{
\item{rankings}{A \code{data.frame} containing the following columns:
\describe{
\item{timepoint}{Integer specifying the timepoint.}
\item{user}{Integer specifying the user id.}
\item{item1, item2,...}{Ranking given to each item.}
}}

\item{preferences}{A \code{data.frame} containing the following columns:
\describe{
\item{timepoint}{Integer specifying the timepoint.}
\item{user}{Integer specifying the user id.}
\item{bottom_item}{Integer identifying the disfavored item.}
\item{top_item}{Integer identifying the preferred item.}
}}
}
\value{
An object to be provided to the \code{data} argument of
\code{\link[=compute_sequentially]{compute_sequentially()}}.
}
\description{
This function is used to prepare and check rank and preference datasets to
ensure they are ready for further processing by \code{\link[=compute_sequentially]{compute_sequentially()}}.
With complete or partial rankings, the \code{rankings} argument must be provided.
With pairwise preference data, the \code{preferences} argument must be provided.
}
\details{
The timepoints, users, and items must all be identified using integers
starting at 1.
}
