% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complete_rankings}
\alias{complete_rankings}
\title{Simulated data with complete rankings}
\format{
A data frame with 400 rows and 12 columns:
\describe{
\item{timepoint}{Integer specifying the timepoint.}
\item{user}{Integer specifying the user id.}
\item{item1, item2,...}{Ranking given to each item.}
}
}
\usage{
complete_rankings
}
\description{
A simulated dataset containing complete rankings of ten items by 400 users,
at 400 separate timepoints. The dataset was simulated using the
\code{sample_mallows} function of the \code{BayesMallows} package
\insertCite{sorensenBayesMallowsPackageBayesian2020}{BayesMallowsSMC2}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
