% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sequentially.R
\name{compute_sequentially}
\alias{compute_sequentially}
\title{Estimate the Bayesian Mallows model sequentially}
\usage{
compute_sequentially(
  data = setup_rank_data(),
  smc_options = set_smc_options(),
  priors = set_priors(),
  model_options = set_model_options()
)
}
\arguments{
\item{data}{An object returned from \code{\link[=setup_rank_data]{setup_rank_data()}}.}

\item{smc_options}{An object returned from \code{\link[=set_smc_options]{set_smc_options()}}.}

\item{priors}{An object returned from \code{\link[=set_priors]{set_priors()}}.}

\item{model_options}{An object returned from \code{\link[=set_model_options]{set_model_options()}}.}
}
\value{
An object of class \code{BayesMallowsSMC2}.
}
\description{
Estimate the Bayesian Mallows model sequentially
}
\references{
\insertRef{chopinSMC2EfficientAlgorithm2013}{BayesMallowsSMC2}

\insertRef{fulopEfficientLearningSimulation2013}{BayesMallowsSMC2}
}
