% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sequentially.R
\name{compute_sequentially}
\alias{compute_sequentially}
\title{Compute the Bayesian Mallows model sequentially}
\usage{
compute_sequentially()
}
\value{
An object
}
\description{
Compute the Bayesian Mallows model sequentially
}
\examples{
compute_sequentially()
}
