% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sequentially.R
\name{compute_sequentially}
\alias{compute_sequentially}
\title{Compute the Bayesian Mallows model sequentially}
\usage{
compute_sequentially(
  data,
  hyperparameters = set_hyperparameters(),
  smc_options = set_smc_options(),
  topological_sorts = NULL
)
}
\arguments{
\item{data}{A dataframe containing partial rankings or pairwise preferences.
If \code{data} contains complete or partial rankings, it must have the following
columns:

\itemize{
\item \code{timepoint}: a numeric vector denoting the timepoint, starting at 1.
\item \code{user}: a vector identifying the user.
\item \code{item1}: ranking of item 1.
\item \code{item2}: ranking of item 2.
\item etc.
}

If data contains pairwise preferences, it must have the following
structure:

\itemize{
\item \code{timepoint}: a numeric vector denoting the timepoint, starting at 1.
\item \code{user}: a vector identifying the user.
\item \code{top_item}: identifier for the preferred item.
\item \code{bottom_item}: identifier for the dispreferred item.
}}

\item{hyperparameters}{A list returned from \code{\link[=set_hyperparameters]{set_hyperparameters()}}.}

\item{smc_options}{A list returned from \code{\link[=set_smc_options]{set_smc_options()}}}

\item{topological_sorts}{A list returned from
\code{\link[=precompute_topological_sorts]{precompute_topological_sorts()}}. Only used with preference data, and
defaults to \code{NULL}.}
}
\value{
An object of class BayesMallowsSMC2.
}
\description{
Compute the Bayesian Mallows model sequentially
}
\examples{
# Compute the model sequentially with complete rankings
mod <- compute_sequentially(
  complete_rankings,
  hyperparameters = set_hyperparameters(n_items = 5),
  smc_options = set_smc_options(n_particles = 100, n_particle_filters = 1)
)

}
