% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precompute_topological_sorts}
\alias{precompute_topological_sorts}
\title{Precompute All Topological Sorts}
\usage{
precompute_topological_sorts(
  prefs,
  n_items,
  output_directory,
  max_files_to_save
)
}
\arguments{
\item{prefs}{An integer matrix (\code{arma::umat}) representing the preference relations.}

\item{n_items}{An integer specifying the number of items to sort.}

\item{output_directory}{A string specifying the directory where the output files will be saved.}

\item{max_files_to_save}{An integer specifying an upper limit on the number of files to save. If
the number of topological sorts exceeds \code{max_files_to_save}, a random sample of size \code{max_files_to_save}
is taken at random, and the corresponding sorts saved.}
}
\value{
This function does not return a value. It performs its operations as a side effect.
}
\description{
This function precomputes all topological sorts for a given preference matrix and
saves them to a specified output directory. It ensures the output directory exists
and creates it if it does not.
}
\details{
The function generates all possible topological sorts for the provided preference matrix
and saves each sort as a binary file in the specified output directory. The output files
are named sequentially as \code{sort0.bin}, \code{sort1.bin}, and so on.
}
