% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyperparameters.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Set hyperparameters}
\usage{
set_hyperparameters(
  n_items,
  alpha_shape = 1,
  alpha_rate = 0.5,
  cluster_concentration = 10,
  n_clusters = 1
)
}
\arguments{
\item{n_items}{Integer defining the number of items.}

\item{alpha_shape}{Shape of gamma prior for alpha.}

\item{alpha_rate}{Rate of gamma prior for alpha.}

\item{cluster_concentration}{Concentration parameter of Dirichlet distribution for cluster probabilities.}

\item{n_clusters}{Integer defining the number of clusters.}
}
\value{
A list
}
\description{
Set hyperparameters
}
