% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc_options.R
\name{set_smc_options}
\alias{set_smc_options}
\title{Set options for SMC2 algorithm}
\usage{
set_smc_options(
  n_particles,
  n_particle_filters,
  resampler = c("multinomial", "residual", "stratified", "systematic"),
  max_rejuvenation_steps = 10,
  verbose = FALSE
)
}
\arguments{
\item{n_particles}{Integer specifying the number of particles used in the
top-level iterated batch importance sampling algorithm.}

\item{n_particle_filters}{Integer specifying the number of particle filters
to run for each top-level particle.}

\item{resampler}{Resampler to use. One of "multinomial", "residual",
"stratified", or "systematic".}

\item{max_rejuvenation_steps}{Maximum number of MCMC steps in rejuvenation
algorithm.}

\item{verbose}{Whether to report progress or not.}
}
\value{
An object to be provided to the \code{smc_options} argument of
\code{\link[=compute_sequentially]{compute_sequentially()}}.
}
\description{
This function is used to set the options for the SMC2 algorithm used to
estimate the Bayesian Mallows model.
}
