% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc_options.R
\name{set_smc_options}
\alias{set_smc_options}
\title{Set SMC options}
\usage{
set_smc_options(
  n_particles = 1000,
  n_particle_filters = 50,
  max_particle_filters = 10000,
  resampling_threshold = n_particles/2,
  metric = "footrule",
  resampler = "multinomial",
  verbose = FALSE,
  trace = FALSE
)
}
\arguments{
\item{n_particles}{Number of particles}

\item{n_particle_filters}{Iniitial number of particle filters for each particle}

\item{max_particle_filters}{Maximum number of particle filters.}

\item{resampling_threshold}{Effective sample size threshold for resampling}

\item{metric}{Metric}

\item{resampler}{resampler}

\item{verbose}{Boolean}

\item{trace}{Logical specifying whether to save static parameters at each
timestep.}
}
\value{
A list
}
\description{
Set SMC options
}
