% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc_options.R
\name{set_smc_options}
\alias{set_smc_options}
\title{Set SMC options}
\usage{
set_smc_options(
  n_particles = 1000,
  n_particle_filters = 50,
  resampling_threshold = n_particles/2,
  metric = "footrule",
  resampler = "multinomial"
)
}
\arguments{
\item{n_particles}{Number of particles}

\item{n_particle_filters}{Number of particle filters for each particle}

\item{resampling_threshold}{Effective sample size threshold for resampling}

\item{metric}{Metric}

\item{resampler}{resampler}
}
\value{
A list
}
\description{
Set SMC options
}
