% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc_options.R
\name{set_smc_options}
\alias{set_smc_options}
\title{Set SMC options}
\usage{
set_smc_options(
  n_particles = 1000,
  n_particle_filters = 50,
  max_particle_filters = 10000,
  resampling_threshold = n_particles/2,
  max_rejuvenation_steps = 20,
  metric = "footrule",
  resampler = "multinomial",
  latent_rank_proposal = "uniform",
  verbose = FALSE,
  trace = FALSE,
  trace_latent = FALSE
)
}
\arguments{
\item{n_particles}{Number of particles}

\item{n_particle_filters}{Initial number of particle filters for each
particle}

\item{max_particle_filters}{Maximum number of particle filters.}

\item{resampling_threshold}{Effective sample size threshold for resampling}

\item{max_rejuvenation_steps}{Maximum number of rejuvenation steps. If the
number of unique particles has not exceeded half the number of particles
after this many steps, the rejuvenation is still stopped.}

\item{metric}{Metric}

\item{resampler}{resampler}

\item{latent_rank_proposal}{latent rank proposal}

\item{verbose}{Boolean}

\item{trace}{Logical specifying whether to save static parameters at each
timestep.}

\item{trace_latent}{Logical specifying whether to sample and save one
complete set of latent rankings for each particle and each timepoint.}
}
\value{
A list
}
\description{
Set SMC options
}
